+++ codalab/server/rest_server.py
@@ -28,7 +28,7 @@ import codalab.rest.legacy
 import codalab.rest.oauth2
 import codalab.rest.titlejs
 import codalab.rest.users
-import codalab.rest.worker
+import codalab.rest.workers
 from codalab.server.authenticated_plugin import UserVerifiedPlugin
 from codalab.server.cookie import CookieAuthenticationPlugin
 from codalab.server.json_api_plugin import JsonApiPlugin
